backend/requirements.txt
flask
python-dotenv
requests
appium-python-client
flask
